brain_app/requirements.txt
django>=4.2,<5.0
onnxruntime>=1.17
numpy
Pillow
matplotlib
reportlab
scipy
gunicorn>=21.2
whitenoise>=6.5
